Create NumbersToWord.CS
Hello 

This assignment has been developed using .NET Core 3.1. Please ensure your environment has the necessary packages installed.

There are a number of assumptions made to this submission, namely: 
There is only a single number within a string, (i.e "Hello 1234" would be valid but "Hello I turn 24 on June 13th" would be invalid)
The inclusion of a currency identifiers such as R,$ are not accepted
The inclusion of numbers equal to or greater than 1 000 000 are not valid. 
A number may be entered with a zero as the first digit, (i.e 012, and 0112129808) and will be valid entry but the first digit will be dropped, as it is a zero.
Any characters that is not a single whitespace within a number will invalidate the input (i.e "1 2 3" will be valid "1   2    3" will be invalid "1,2 3" will be invalid)